Make the SwarmEvent report everything (#1515)
* Improve the SwarmEvent to report everything

* Address review

* Update swarm/src/lib.rs

+++ core/tests/network_dial_error.rs
@@ -258,7 +258,7 @@ fn multiple_addresses_err() {
                         assert_eq!(attempts_remaining, 0);
                         return Poll::Ready(Ok(()));
                     } else {
-                        assert_eq!(attempts_remaining, addresses.len());
+                        assert_eq!(attempts_remaining, addresses.len() as u32);
                     }
                 },
                 Poll::Ready(_) => unreachable!(),
